@@ -1,13 +1,20 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <IsPackable>False</IsPackable>
-    <Version>2.2.0</Version>
+    <Version>2.9.0-alpha3</Version>
+    <AspNetCoreHostingModel>OutOfProcess</AspNetCoreHostingModel>
     <ServerGarbageCollection>false</ServerGarbageCollection>
     <RetainVMGarbageCollection>false</RetainVMGarbageCollection>
     <TargetLatestRuntimePatch>true</TargetLatestRuntimePatch>
+    <PublishReadyToRun>true</PublishReadyToRun>
+    <PublishReadyToRunShowWarnings>true</PublishReadyToRunShowWarnings>
+    <PublishTrimmed>true</PublishTrimmed>
     <NoWarn>FS2003</NoWarn>
   </PropertyGroup>
+  <ItemGroup>
+    <TrimmerRootAssembly Include="System.Threading.Tasks" />
+  </ItemGroup>
   <ItemGroup>
     <Compile Include="GCMiddleware.fs" />
     <Compile Include="Middleware.fs" />
