@@ -13,9 +13,9 @@ open Newtonsoft.Json
 open Newtonsoft.Json.Serialization
 open System
 open System.Threading
-open NCoreUtils.Images.Optimization
+// open NCoreUtils.Images.Optimization
 
-type Startup (env : IHostingEnvironment) =
+type Startup (env : IWebHostEnvironment) =
 
   static let configureSourceExtractors =
     Action<CompositeImageSourceExtractorBuilder>
@@ -47,14 +47,14 @@ type Startup (env : IHostingEnvironment) =
         b.ClearProviders()
           .SetMinimumLevel(LogLevel.Information)
           |> ignore
-        match env.IsDevelopment () with
-        | true -> b.AddConsole () |> ignore
-        | _    -> b.AddGoogleSink (configuration.GetSection "Google") |> ignore
+        match env.EnvironmentName with
+        | "Development" -> b.AddConsole () |> ignore
+        | _             -> b.AddGoogleSink (configuration.GetSection "Google") |> ignore
       )
       .AddPrePopulatedLoggingContext()
       .AddSingleton(JsonSerializerSettings (ReferenceLoopHandling = ReferenceLoopHandling.Ignore, ContractResolver = CamelCasePropertyNamesContractResolver ()))
       .AddSingleton<IHttpContextAccessor, HttpContextAccessor>()
-      .AddSingletonImageOptimization<JpegoptimOptimization>()
+      // .AddSingletonImageOptimization<JpegoptimOptimization>()
       .AddImageSourceExtractors(configureSourceExtractors)
       .AddImageDestinationExtractors(configureDestinationExtractors)
       .AddImageMagickResizer()
@@ -63,9 +63,10 @@ type Startup (env : IHostingEnvironment) =
   member __.Configure (app: IApplicationBuilder, serviceConfiguration : ServiceConfiguration, loggerFactory : ILoggerFactory) =
     let semaphore = new SemaphoreSlim (serviceConfiguration.MaxConcurrentOps, serviceConfiguration.MaxConcurrentOps)
 
-    let logger = loggerFactory.CreateLogger typeof<ImageMagick.MagickNET>
-    ImageMagick.MagickNET.SetLogEvents (ImageMagick.LogEvents.Image ||| ImageMagick.LogEvents.Coder ||| ImageMagick.LogEvents.Transform)
-    ImageMagick.MagickNET.Log.AddHandler (fun _ e -> logger.LogInformation (sprintf "[%s] %s" (e.EventType.ToString ()) e.Message))
+    if "Development" = env.EnvironmentName then
+      let logger = loggerFactory.CreateLogger typeof<ImageMagick.MagickNET>
+      ImageMagick.MagickNET.SetLogEvents (ImageMagick.LogEvents.Image ||| ImageMagick.LogEvents.Coder ||| ImageMagick.LogEvents.Transform)
+      ImageMagick.MagickNET.Log.AddHandler (fun _ e -> logger.LogInformation (sprintf "[%s] %s" (e.EventType.ToString ()) e.Message))
 
     app
       .UsePrePopulateLoggingContext()
